requirements.txt
fastapi
uvicorn
gensim
PyMuPDF
python-magic
joblib
tensorflow
Flask-Cors
python-multipart